Move e/d command back to mod domain, now that we made space
